requirements.txt
streamlit
openai-whisper
soundfile
numpy
python-docx
